requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-environ==0.4.5
django-filter==2.2.0
django-guardian==2.1.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
Markdown==3.1.1
opencc-python-reimplemented==0.1.5
pycryptodome==3.9.0
Pygments==2.4.2
PyJWT==1.7.1
python-weixin==0.5.0
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
xmltodict==0.12.0
